requirements.txt
pymongo[srv]==4.2.0
pandas
numpy
matplotlib
notebook
PyYAML
seaborn
sklearn
flask
six
-e .